+++ mandatory/src/strlen_char.c
@@ -6,15 +6,15 @@
 /*   By: alafdili <alafdili@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/01 15:03:09 by alafdili          #+#    #+#             */
-/*   Updated: 2024/03/15 23:53:41 by alafdili         ###   ########.fr       */
+/*   Updated: 2024/04/14 14:17:05 by alafdili         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "so_long.h"
 
-size_t	strlen_char(char *s)
+int	strlen_char(char *s)
 {
-	size_t	counter;
+	int	counter;
 
 	counter = 0;
 	while (s[counter] && s[counter] != '\n')
